+++ webview/util.py
@@ -17,6 +17,10 @@ from .js import api, npo
 default_html = '<!doctype html><html><head></head><body></body></html>'
 
 
+class WebViewException(Exception):
+    pass
+
+
 def base_uri(relative_path=''):
     """ Get absolute path to resource, works for dev and for PyInstaller """
     try:
@@ -61,10 +65,8 @@ def parse_file_type(file_type):
 
 
 def parse_api_js(api_instance):
-    print(0)
     func_list = [str(f) for f in dir(api_instance) if callable(getattr(api_instance, f)) and str(f)[0] != '_']
     js_code = npo.src + api.src % func_list
-    print(api_instance)
     return js_code
